aligning of navbar links, button styles, structured file structures, linked profiles on home page
