--- src/components/Styles.css
@@ -1,228 +1,0 @@
-.allText {
-  font-family: "Times New Roman", Times, serif;
-  font-size: 1em;
-}
-
-.LoginButton :hover {
-  background-color:#7e191b;
-}
-
-.LoginButton{
-  color:  white;
-  border: white;
-  border-radius: 0.1em;
-  background-color: #c21a09;
-  font-size: large;
-  padding: 0.5em;
-}
-
-.HeaderFooterStyle {
-  background-color: #c21a09;
-  color: white;
-  font-family: "Times New Roman", Times, serif;
-  text-align: left;
-  padding: 0.5em;
-  margin-top: 0em;
-  font-size: 1em;
-
-}
-
-.photoStyle {
-  margin-top: auto;
-  padding: 0.5em;
-  align-items: center;
-}
-
-.NavOptions {
-  text-align: right;
-}
-.NavStyle {
-  /*overflow: hidden;*/
-  background-color: #7e191b;
-  font-size: medium;
-  font-weight: bold;
-  color: #ffffff;
-  padding: 10px 10px;
-  text-decoration: none;
-  display: block;
-}
-
-/* Style the NavStyle links */
-.NavStyle a {
-  float: left;
-  display: block;
-  color: #ffffff;
-  text-align: center;
-  padding: 10px 10px;
-  text-decoration: none;
-}
-
-.NavStyle a:visited {
-  color: #ffffff;
-}
-.NavStyle a:active {
-  color: #ffffff;
-}
-
-/* Change color on hover */
-.top_nav a:hover {
-  background-color: #c21a09;
-  color: #ffffff;
-}
-
-/*-----------------------------*/
-/* Buttons on CreateAccount.js */
-
-
-
-.resetButton {
-  background-color: #c21a09;
-  color: #ffffff;
-  border: 0ch;
-  text-decoration: none;
-  flex: 1; 
-  flex-direction: row;
-  justify-content: space-around;
-  font-size: medium;
-  padding: 12px 20px;
-  margin: 5px 10px 0px 10px;
-}
-
-/*
-margin: 4px 2px;
-padding: 12px 20px;
-cursor: pointer;
-
-.resetButton :hover{
-  background-color: #7e191b;
-  color: #ffffff;
-}
-
-.buttonStyle {
-  padding: 5;
-  flex: 1; 
-  flex-direction: row;
-  justify-content: space-around;
-}
-*/
-
-.submitButton {
-  background-color: #c21a09;
-  color: #ffffff;
-  border: 0ch;
-  padding: 12px 20px;
-  text-decoration: none;
-  cursor: pointer;
-  font-size: medium;
-}
-
-/*
-.submitButton :hover{
-  background-color: #7e191b;
-  color: #ffffff;
-}
-*/
-
-/* Buttons on CreateAccount.js */
-/*-----------------------------*/
-
-.website-ul {
-  list-style-type: none;
-  margin: 0;
-}
-
-.menu-ul {
-  list-style-type: none;
-  margin: 0;
-  float: right;
-  margin-right: 20px;
-  margin-top: 20px;
-}
-
-.menu-ul li {
-  display: inline;
-  margin-left: 20px;
-  margin-right: 20px;
-}
-
-.logoStyle{
-  width: 320px;
-  height: 90px;
-  padding: 0.1em;
-}
-
-.menuStyle{ 
-  width: 100px;
-  height: 80px;
-}
-
-.header {
-  background-color: #c21a09;
-  color: white;
-  font-family: "Times New Roman", Times, serif;
-  padding: 0.2em;
-  font-size: 1.5em;
-}
-
-.header :any-link {
-  color: #ffffff;
-  text-decoration: none;
-}
-
-.header :hover {
-  /*color: #7e191b; */
-  color: #380000;
-}
-
-.brandNameFontStyle {
-  color: #ffffff;
-  /* font-family: initial; */
-  font-weight: 400;
-  font-size: large;
-}
-
-.navOptions{
-  color: #ffffff;
-  font-weight: 300;
-  font-size: medium;
-}
-
-.navSpacing{
-  margin-left:30px;
-  margin-right:30px;
-}
-
-.navOptions :hover{
-  text-decoration: underline;
-  /* zoom: 1.05; */
-}
-
-.navBarColorStyle {
-  background-color:#c21a09;
-}
-
-.alignCenterForm {
-  margin-left: 40%;
-  margin-top: 10%;
-} 
-
-/* createAccount input area style */
-.inputAreaStyle {
-  padding: 0.2cm;
-  width: 7cm;
-}
-
-/* .Comment {
-  margin-left: 35%;
-  margin-top: 10%;
-} */
-
-.Comment {
-  min-height: 75vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 1vmin);
-  color: black;
-}
