+++ src/components/pages/Footer.js
@@ -4,7 +4,7 @@ import './Styles.css';
 export default class Footer extends React.Component{
     render(){
         return (
-            <div className="HeaderFooterStyle"> 
+            <div className="FooterStyle"> 
                 <h1>About us</h1>
             </div>
         );
